+++ js/_loader.js
@@ -1,3 +1,5 @@
+/* globals module, bfg_script_srcs, bfg_scripts_loading, bfg_scripts_loaded, bfg_scripts_events_queue */
+
 // Make sure all globals are set, but don't overwrite if they are
 window.bfg_script_srcs = window.bfg_script_srcs || {};
 window.bfg_scripts_loading = window.bfg_scripts_loading || [];