+++ js/scripts.js
@@ -1,4 +1,4 @@
-import im from '../node_modules/include-media-export/include-media.js';
+// import im from '../node_modules/include-media-export/include-media.js';
 import fitvids from '../node_modules/fitvids/dist/fitvids.js';
 
 import './_svgs.js';